refactor: move addGoal to utils
