+++ frontend/src/hooks/useGoals.ts
@@ -1,0 +1,1 @@
+export const useGoal = () => {};
